src/main/java/org/httpserver/handler/HtmlHandler.java
package org.httpserver.handler;

import org.httpserver.request.Request;
import org.httpserver.response.Response;
import org.httpserver.response.ResponseBuilder;
import org.httpserver.response.StatusCode;
import org.httpserver.server.HttpMethod;

import java.util.List;

public class HtmlHandler implements Handler{

    public List<String> allowedHttpMethods() {
        return List.of(HttpMethod.GET.getHttpMethod());
    }

    @Override
    public Response handleResponse(Request request) {
        return new ResponseBuilder()
                .withStatusCode(StatusCode.OK)
                .withHeaderName("Content-Type")
                .withHeaderValue("text/html;charset=utf-8")
                .withBody("<html><body><p>HTML Response</p></body></html>")
                .build();
    }
}
